requirements.txt
ipywidgets
matplotlib
numpy
notebook
pyyaml
scikit_learn
scipy
seaborn
setuptools
torch
torchdiffeq
torchvision
tqdm
wandb
hydra-core 
hydra-joblib-launcher 
hydra-submitit-launcher
pytest-cov
build
charset-normalizer